requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-jet==1.0.8
django-phonenumber-field==5.0.0
django-twilio==0.13.0.post1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
openpyxl==3.0.5
phonenumbers==8.12.11
Pillow==8.0.0
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2020.1
qrcode==6.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
twilio==6.63.2
urllib3==1.25.10
whitenoise==5.2.0
